requirements.txt
jira>=1.0.10
pyyaml
prompter
python-editor
attrs
prettytable
cmd2
iso8601
six